\name{comparison}
\alias{comparison}
\title{Compares the output of the optimizaton for different initialization values.}
\usage{
  comparison(estimation)
}
\arguments{
  \item{estimation}{An rds2 estimation object}
}
\value{
  Used only for printing
}
\description{
  Compares the output of the optimizaton for different
  initialization values.
}
\author{
  johnros
}

