\name{estimate.rds}
\alias{estimate.rds}
\title{Main function in rds2 package. Returns the ML estimate of population size and degree distribution.}
\usage{
  estimate.rds(data, init, const, arc, maxit = 20000, ...)
}
\arguments{
  \item{data}{}

  \item{init}{}

  \item{const}{}

  \item{arc}{}

  \item{maxit}{}

  \item{...}{}
}
\value{
  TBC
}
\description{
  Main function in rds2 package. Returns the ML estimate of
  population size and degree distribution.
}
\author{
  Jonathan Rosenblatt
}

