\name{inv.qnorm.theta}
\alias{inv.qnorm.theta}
\title{Inverts the map of theta to the real line.}
\usage{
  inv.qnorm.theta(qnorm.theta, const, range = 20,
    minimum = -10)
}
\arguments{
  \item{qnorm.theta}{}

  \item{const}{}

  \item{range}{}

  \item{minimum}{}
}
\value{
  TBC
}
\description{
  Inverts the map of theta to the real line.
}
\author{
  johnros
}

