\name{make.Sij}
\alias{make.Sij}
\title{Make the Sij matrix assuming no droouts from the snowball}
\usage{
  make.Sij(data)
}
\arguments{
  \item{data}{ An integer vector of the degrees of each individual sampled.}
}
\value{
  TBC
}
\description{
  Make the Sij matrix assuming no droouts from the snowball
}
\author{
  johnros
}

