\name{qnorm.theta}
\alias{qnorm.theta}
\title{Maps the theta parameter to the real line.}
\usage{
  qnorm.theta(theta, const, range = 20, minimum = -10)
}
\arguments{
  \item{theta}{}

  \item{const}{Controls the}

  \item{range}{}

  \item{minimum}{}
}
\value{
  TBC
}
\description{
  Maps the theta parameter to the real line.
}
\author{
  johnros
}

