\docType{package}
\name{rds2-package}
\alias{rds2}
\alias{rds2-package}
\title{Population size estimation for despondant driven sampling}
\description{
  Estimates oppulation size and degree distribution in
  respondant driven samples(RDS).
}
\details{
  \tabular{ll}{ Package: \tab rds2\cr Type: \tab Package\cr
  Version: \tab 0.5\cr Date: \tab 2011-12-25\cr License:
  \tab GPL (>= 2)\cr }

  Maximum likelihood estimation of population size using
  the methodology in [cite].
}
\author{
  Jonathan Rosenblatt
}

