% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/serialize.R
\name{serialize}
\alias{deserialize}
\alias{serialize}
\title{Serialize or deserialize an R object according to its RDS2 serialization.}
\usage{
serialize(object)

deserialize(object)
}
\arguments{
\item{object}{ANY. The R object to serialize.}
}
\value{
For serialize, the serialized R object. For deserialize, the
  deserialized R object.

  The function \code{attr(object, "RDS2.serialize")$write} will be
  used to perform the serialization and the
  \code{attr(object, "RDS2.serialize")$read} function will be used
  to perform the deserialization.
}
\description{
Serialize or deserialize an R object according to its RDS2 serialization.
}
\seealso{
\code{\link{saveRDS}}
}

