\name{sum.ranks}
\alias{sum.ranks}
\title{Utility functions for creating snowball matrix.}
\usage{
  sum.ranks(data)
}
\arguments{
  \item{data}{The sample}
}
\value{
  The sum of degrees for each rank
}
\description{
  Utility functions for creating snowball matrix.
}
\author{
  Jonathan Rosnblatt
}

