\name{compute.S}
\alias{compute.S}
\title{Compares the output of the optimizaton for different initialization values.}
\usage{
  compute.S(Sij)
}
\arguments{
  \item{Sij}{ An integer matrix of the counts of individuals with rank j at sampling period t. Row names correspond to the degrees.}
}
\value{
  Used for creating the Snowball size given an Sij matrix
}
\description{
  Compares the output of the optimizaton for different
  initialization values.
}
\author{
  johnros
}

