\name{estimate.rds2}
\alias{estimate.rds2}
\title{Main function in rds2 package. Returns the ML estimate of population size and degree distribution.}
\usage{
  estimate.rds2(data, Sij, init, const, arc, maxit = 10000,
    theta, ...)
}
\arguments{
  \item{data}{}

  \item{Sij}{}

  \item{init}{}

  \item{const}{}

  \item{arc}{}

  \item{maxit}{}

  \item{theta}{}

  \item{...}{}
}
\value{
  TBC
}
\description{
  Main function in rds2 package. Returns the ML estimate of
  population size and degree distribution.
}
\examples{
rnorm(1)
}
\author{
  johnros
}

