\name{sum.ranks}
\alias{sum.ranks}
\title{Utility functions for creating snowball matrix.}
\usage{
  sum.ranks(data)
}
\arguments{
  \item{data}{ An integer vector of the degrees of each individual sampled.}
}
\value{
  The sum of degrees for each rank
}
\description{
  Utility functions for creating snowball matrix.
}
\author{
  Jonathan D. Rosenblatt \email{john.ros@gmail.com}
}

